cleanup: fix 22 miscellaneous clippy warnings
- match→if let, for_kv_map→keys(), needless_range_loop→enumerate
- manual_clamp, manual_is_multiple_of, manual_range_contains
- single_char_add_str, unnecessary_cast, collapsible_if
- missing_transmute_annotations, deprecated into_path→keep
- Rename yaml_spec::from_str to parse to avoid trait confusion
- Add Default impl for MacOSSandbox
- Add #[allow] for intentional enum_variant_names

+++ core/src/memory/dreaming.rs
@@ -413,7 +413,7 @@ impl DreamDaemon {
 
         report.graph_decay = self.graph_store.apply_decay(&self.graph_decay).await?;
         let decayed_count = self.database.apply_fact_decay(
-            self.memory_decay.half_life_days as f32,
+            self.memory_decay.half_life_days,
             self.memory_decay.min_strength,
         ).await?;
         report.memory_decay = MemoryDecayReport {
